🎨

+++ assets-catalog/src/main.py
@@ -14,18 +14,21 @@ class Upload_asset:
         self._key_lock = threading.Lock()
 
     def upload(self, asset):
-        self._key_lock.acquire()
-        if is_exist(asset):
-            self._response = 'Error: The file is exists!'
+        if os.path.exists(watch_directory):
+            self._key_lock.acquire()
+            if is_exist(asset):
+                self._response = 'Error: The file is exists!'
+            else:
+                self._response = upload_file(asset)
+            self._key_lock.release()
         else:
-            self._response = upload_file(asset)
-        self._key_lock.release()
+            self._response = 'Error: The watch directory not found!'
 
 
 def upload_file(asset):
     try:
-       if write_to_cash(asset) & upload_to_server(asset):
-          return 'Upload file is success!'
+        if write_to_cash(asset) & upload_to_server(asset):
+            return 'Upload file is success!'
     except:
         return 'Upload file is failed!'
 
@@ -34,7 +37,7 @@ def write_to_cash(asset):
     try:
         with open(watch_directory) as data_file:
             asset_data = json.load(data_file)
-        asset_data.append({"filename":asset["filename"]})
+        asset_data.append({"filename": asset["filename"]})
         with open(watch_directory, 'w') as json_file:
             json.dump(asset_data, json_file,
                       indent=4,
@@ -45,7 +48,7 @@ def write_to_cash(asset):
 
 
 def upload_to_server(asset):
-    try:   
+    try:
         if asset['filename']:
             fn = os.path.basename(asset['filename'])
             open(fn, 'w').write(asset['content'])
